backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Cors==4.0.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
passlib==1.7.4
mysql-connector-python==9.1.0
colorama==0.4.6
